Antene

+++ AntenneLED.py
@@ -92,7 +92,7 @@ def Frequenz(Distanz):
 
 def LEDAntenne():                                           #Wie viele LEDs an der Antenne gehen an
     global LEDAntenne
-    LEDAntenne = int(round((Tonindex-LowTon+1)*(LED_COUNT_2/(HighTon-LowTon+1)),0))
+    LEDAntenne = round((Tonindex-LowTon+1)*(LED_COUNT_2/(HighTon-LowTon+1)),0)
 
 
 def showColor(strip, color):                     #LED Streifen an machen in color
